chore: node adapters
