+++ node/adapters/errors.go
@@ -1,0 +1,9 @@
+package adapters
+
+import (
+	"errors"
+)
+
+var (
+	ErrMessengerNotReady = errors.New("messenger not ready")
+)
